renamed src folder to project name
