+++ adsorpyon/Properties/CH4.prop
