+++ adsorpyon/Properties/CO2.prop
