+++ adsorpyon/Properties/N2.prop
