+++ adsorpyon/Properties/NH3.prop
