+++ adsorpyon/Properties/methanol.prop
